backend/requirements.txt
Flask==2.2.2
boto3==1.26.0
python-dotenv==0.19.2
Werkzeug==2.2.2
